Update index.md
Hi there! Nice to e-meet you, 
I am Polina From the NOWNodes team.
NOWNodes is a blockchain-as-a-service solution that lets users get access to full Nodes and blockbook Explorers via API. Instant access to 56 blockchain networks, including BTC, ETH, and BSC, which amount has over 1000 coins and tokens.
We would like to be listed in this article
Could we discuss the terms and conditions of cooperation?

Our features:
- Access to 50+ blockchain nodes
- Free API keys 
- Block Explorer
- Shared, dedicated,  archive and testnet nodes 
- Backup solution 
- API Response Time ⩽ 1 sec
- 24/7 support team 
- Personal account manager 

Thanks in advance! Looking forward to hearing you soon!